requirements.txt
groq
seleniumbase
python-dotenv